+++ test/tests_project_loading.jl
@@ -97,8 +97,10 @@ end
         test_load_custom_medium_categories()
     end
 
+    include("initialization/bus.jl")
+
     include("order_of_operations/bus_output_priorities.jl")
     include("order_of_operations/bus_to_bus.jl")
     include("order_of_operations/heat_pumps_wrong.jl")
-    include("order_of_operations/project_loading_switch.jl")
+    include("order_of_operations/storage_loading_switch.jl")
 end